Create a new admonition for version
- VERSION to indicate version info inside sections. When further
  descriptions are needed, authors can use this admonition.
- Create style directory and put config files in it.
- Adjust Makefile
