--- ditaa.cfg
@@ -1,1 +1,0 @@
-skinparam defaultFontName  monochrome true
