+++ style/ditaa.cfg
@@ -1,0 +1,1 @@
+skinparam defaultFontName  monochrome true
